% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_di.R
\name{compute_dissim_di}
\alias{compute_dissim_di}
\title{Compute Functional Diversity Using All Dissimilarities}
\usage{
compute_dissim_di(trait_dissim_list, site_sp_df, trait_comb_df)
}
\arguments{
\item{trait_dissim_list}{[\verb{list(list(matrix(1+)))}]\cr{}
a list containing list of dissimilarity matrices
with their names}

\item{site_sp_df}{[\code{data.frame()}]\cr{}
a tidy species data.frame with a row defining a
species in a site with its abundance}

\item{trait_comb_df}{[\code{data.frame()}]\cr{}
a data.frame with the list of trait combinations
considered as well as the number of traits included}

\item{var_type}{[\code{character(1)}]\cr{}
a string giving the column on which \code{site_sp_df}
should be spread}
}
\description{
Compute Functional Diversity Using All Dissimilarities
}
