% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_other_fd.R
\name{compute_other_fd}
\alias{compute_other_fd}
\title{Compute FD indices without abundances}
\usage{
compute_other_fd(site_sp_df, trait_comb, trait_df, var_type = "env")
}
\arguments{
\item{site_sp_df}{[\code{data.frame()}]\cr{}
a tidy species data.frame with a row defining a
species in a site with its abundance}

\item{trait_comb}{[\verb{list(character(1+))}]\cr{}
a list contaning combinations of trait combinations}

\item{var_type}{[\code{character(1)}]\cr{}
a string giving the column on which \code{site_sp_df}
should be spread}
}
\description{
Wrapper around \code{\link[FD:dbFD]{FD::dbFD()}}
}
