% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_trait_dissim.R
\name{compute_trait_dissim}
\alias{compute_trait_dissim}
\title{Compute Trait Dissimilarities}
\usage{
compute_trait_dissim(trait_comb_list, trait_df)
}
\arguments{
\item{trait_comb_list}{[\verb{list(character(1+))}]\cr{}
a list of character vectors containing the trait
combination on which dissimilarity should be computed}

\item{trait_df}{[\code{data.frame()}]\cr{}
a trait data.frame with species as rows and traits as
columns}
}
\value{
a list of dissimilarity matrices with concatenated names from
trait combinations
}
\description{
For each provided trait combination in a list, returns a list of
dissimilarity matrices computed using \code{funrar}. This function computes
euclidean dissimilarity matrices
}
