% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_traits.R
\name{generate_traits}
\alias{generate_traits}
\title{Generate uncorrelated traits}
\usage{
generate_traits(n_species, n_traits)
}
\arguments{
\item{n_species}{[\code{integer(1)}]\cr{}
the number of species}

\item{n_traits}{[\code{integer(1)}]\cr{}
the number of traits}
}
\value{
a matrix with \code{n_species} rows and \code{n_traits} column with uniform
uncorrelated traits between 0 and 1
}
\description{
Generate uncorrelated traits
}
\examples{
generate_traits(3, 2)

}
