% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_env_fd_obs_ses.R
\name{plot_env_fd_obs_ses_two}
\alias{plot_env_fd_obs_ses_two}
\title{Figure of FD vs. Environment Comparing Observed and SES values}
\usage{
plot_env_fd_obs_ses_two(full_fd_df, plot_labs = "AUTO")
}
\arguments{
\item{full_fd_df}{[\code{data.frame()}]\cr{}
data.frame containing both observed and SES values of
functional diversity}
}
\description{
Figure of FD vs. Environment Comparing Observed and SES values
}
