% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_zero_one.R
\name{scale_zero_one}
\alias{scale_zero_one}
\title{Scale Vector between Zero and One}
\usage{
scale_zero_one(vec)
}
\arguments{
\item{vec}{a numeric vector}
}
\description{
With a numeric vector scale between zero and one.
}
\examples{

vec = stats::rnorm(1000)

zero_one = scale_zero_one(vec)

range(vec)
range(zero_one)

}
